requirements.txt
gym~=0.15.4
numpy~=1.18.4
pyzmq~=19.0.1
PyYAML~=5.3.1
dataclasses~=0.6
opencv-python~=4.2.0.34
carnivalmirror~=0.6.2
torch~=1.5.0
torchvision~=0.6.0
tqdm~=4.46.0
Pillow~=7.0.0
setuptools~=45.2.0
pyglet~=1.3.2